Refactor: extract widgets & restructuring
